fix namespace for providers that were moved
